+++ lib/autoupdate/agent/config.go
@@ -19,98 +19,187 @@
 package agent
 
 import (
-	"context"
-	"log/slog"
+	"errors"
+	"io/fs"
 	"os"
-	"path/filepath"
-	"text/template"
+	"strings"
+	"time"
 
 	"github.com/google/renameio/v2"
 	"github.com/gravitational/trace"
+	"gopkg.in/yaml.v3"
 )
 
 const (
-	updateServiceTemplate = `# teleport-update
-[Unit]
-Description=Teleport auto-update service
+	// updateConfigName specifies the name of the file inside versionsDirName containing configuration for the teleport update.
+	updateConfigName = "update.yaml"
 
-[Service]
-Type=oneshot
-ExecStart={{.LinkDir}}/bin/teleport-update update
-`
-	updateTimerTemplate = `# teleport-update
-[Unit]
-Description=Teleport auto-update timer unit
-
-[Timer]
-OnActiveSec=1m
-OnUnitActiveSec=5m
-RandomizedDelaySec=1m
-
-[Install]
-WantedBy=teleport.service
-`
+	// UpdateConfig metadata
+	updateConfigVersion = "v1"
+	updateConfigKind    = "update_config"
 )
 
-// Setup installs service and timer files for the teleport-update binary.
-// Afterwords, Setup reloads systemd and enables the timer with --now.
-func Setup(ctx context.Context, log *slog.Logger, linkDir, dataDir string) error {
-	err := writeConfigFiles(linkDir, dataDir)
-	if err != nil {
-		return trace.Errorf("failed to write teleport-update systemd config files: %w", err)
-	}
-	svc := &SystemdService{
-		ServiceName: "teleport-update.timer",
-		Log:         log,
-	}
-	if err := svc.Sync(ctx); err != nil {
-		return trace.Errorf("failed to sync systemd config: %w", err)
-	}
-	if err := svc.Enable(ctx, true); err != nil {
-		return trace.Errorf("failed to enable teleport-update systemd timer: %w", err)
-	}
-	return nil
+// UpdateConfig describes the update.yaml file schema.
+type UpdateConfig struct {
+	// Version of the configuration file
+	Version string `yaml:"version"`
+	// Kind of configuration file (always "update_config")
+	Kind string `yaml:"kind"`
+	// Spec contains user-specified configuration.
+	Spec UpdateSpec `yaml:"spec"`
+	// Status contains state configuration.
+	Status UpdateStatus `yaml:"status"`
 }
 
-func writeConfigFiles(linkDir, dataDir string) error {
-	servicePath := filepath.Join(linkDir, serviceDir, updateServiceName)
-	err := writeTemplate(servicePath, updateServiceTemplate, linkDir, dataDir)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	timerPath := filepath.Join(linkDir, serviceDir, updateTimerName)
-	err = writeTemplate(timerPath, updateTimerTemplate, linkDir, dataDir)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	return nil
+// UpdateSpec describes the spec field in update.yaml.
+type UpdateSpec struct {
+	// Proxy address
+	Proxy string `yaml:"proxy"`
+	// Group specifies the update group identifier for the agent.
+	Group string `yaml:"group"`
+	// URLTemplate for the Teleport tgz download URL.
+	URLTemplate string `yaml:"url_template"`
+	// Enabled controls whether auto-updates are enabled.
+	Enabled bool `yaml:"enabled"`
+	// Pinned controls whether the active_version is pinned.
+	Pinned bool `yaml:"pinned"`
 }
 
-func writeTemplate(path, t, linkDir, dataDir string) error {
-	dir, file := filepath.Split(path)
-	if err := os.MkdirAll(dir, systemDirMode); err != nil {
-		return trace.Wrap(err)
+// UpdateStatus describes the status field in update.yaml.
+type UpdateStatus struct {
+	// ActiveVersion is the currently active Teleport version.
+	ActiveVersion string `yaml:"active_version"`
+	// BackupVersion is the last working version of Teleport.
+	BackupVersion string `yaml:"backup_version"`
+	// SkipVersion is the last reverted version of Teleport.
+	SkipVersion string `yaml:"skip_version"`
+}
+
+// readConfig reads UpdateConfig from a file.
+func readConfig(path string) (*UpdateConfig, error) {
+	f, err := os.Open(path)
+	if errors.Is(err, fs.ErrNotExist) {
+		return &UpdateConfig{
+			Version: updateConfigVersion,
+			Kind:    updateConfigKind,
+		}, nil
 	}
+	if err != nil {
+		return nil, trace.Errorf("failed to open: %w", err)
+	}
+	defer f.Close()
+	var cfg UpdateConfig
+	if err := yaml.NewDecoder(f).Decode(&cfg); err != nil {
+		return nil, trace.Errorf("failed to parse: %w", err)
+	}
+	if k := cfg.Kind; k != updateConfigKind {
+		return nil, trace.Errorf("invalid kind %q", k)
+	}
+	if v := cfg.Version; v != updateConfigVersion {
+		return nil, trace.Errorf("invalid version %q", v)
+	}
+	return &cfg, nil
+}
+
+// writeConfig writes UpdateConfig to a file atomically, ensuring the file cannot be corrupted.
+func writeConfig(filename string, cfg *UpdateConfig) error {
 	opts := []renameio.Option{
 		renameio.WithPermissions(configFileMode),
 		renameio.WithExistingPermissions(),
 	}
-	f, err := renameio.NewPendingFile(path, opts...)
+	t, err := renameio.NewPendingFile(filename, opts...)
 	if err != nil {
 		return trace.Wrap(err)
 	}
-	defer f.Cleanup()
+	defer t.Cleanup()
+	err = yaml.NewEncoder(t).Encode(cfg)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	return trace.Wrap(t.CloseAtomicallyReplace())
+}
 
-	tmpl, err := template.New(file).Parse(t)
-	if err != nil {
-		return trace.Wrap(err)
+func validateConfigSpec(spec *UpdateSpec, override OverrideConfig) error {
+	if override.Proxy != "" {
+		spec.Proxy = override.Proxy
 	}
-	err = tmpl.Execute(f, struct {
-		LinkDir string
-		DataDir string
-	}{linkDir, dataDir})
-	if err != nil {
-		return trace.Wrap(err)
+	if override.Group != "" {
+		spec.Group = override.Group
 	}
-	return trace.Wrap(f.CloseAtomicallyReplace())
+	switch override.URLTemplate {
+	case "":
+	case "default":
+		spec.URLTemplate = ""
+	default:
+		spec.URLTemplate = override.URLTemplate
+	}
+	if spec.URLTemplate != "" &&
+		!strings.HasPrefix(strings.ToLower(spec.URLTemplate), "https://") {
+		return trace.Errorf("Teleport download URL must use TLS (https://)")
+	}
+	if override.Enabled {
+		spec.Enabled = true
+	}
+	if override.Pinned {
+		spec.Pinned = true
+	}
+	return nil
+}
+
+// Status of the agent auto-updates system.
+type Status struct {
+	UpdateSpec   `yaml:",inline"`
+	UpdateStatus `yaml:",inline"`
+	FindResp     `yaml:",inline"`
+}
+
+// FindResp summarizes the auto-update status response from cluster.
+type FindResp struct {
+	// Version of Teleport to install
+	TargetVersion string `yaml:"target_version"`
+	// Flags describing the edition of Teleport
+	Flags InstallFlags `yaml:"flags"`
+	// InWindow is true when the install should happen now.
+	InWindow bool `yaml:"active"`
+	// Jitter duration before an automated install
+	Jitter time.Duration `yaml:"jitter"`
+}
+
+// InstallFlags sets flags for the Teleport installation
+type InstallFlags int
+
+const (
+	// FlagEnterprise installs enterprise Teleport
+	FlagEnterprise InstallFlags = 1 << iota
+	// FlagFIPS installs FIPS Teleport
+	FlagFIPS
+)
+
+func (i InstallFlags) MarshalYAML() (any, error) {
+	return i.Strings(), nil
+}
+
+func (i InstallFlags) Strings() []string {
+	var out []string
+	for _, flag := range []InstallFlags{
+		FlagEnterprise,
+		FlagFIPS,
+	} {
+		if i&flag != 0 {
+			out = append(out, flag.String())
+		}
+	}
+	return out
+}
+
+func (i InstallFlags) String() string {
+	switch i {
+	case 0:
+		return ""
+	case FlagEnterprise:
+		return "Enterprise"
+	case FlagFIPS:
+		return "FIPS"
+	}
+	return "Unknown"
 }
