+++ lib/autoupdate/agent/setup_test.go
